more silence warnings when using -Wconversion
git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/trunk@2588 c368ceb8-fd28-0410-a457-e46ee1d7cd4a

+++ src/lib/nonblocking.c
@@ -944,8 +944,8 @@ err_check:
 
         /* if at least one process has a non-zero request, all processes must
          * participate the collective read/write */
-        do_read    = do_io[0];
-        do_write   = do_io[1];
+        do_read    = (int)do_io[0];
+        do_write   = (int)do_io[1];
         newnumrecs = do_io[3];
     }
     else {
