+++ src/lib/var.c
@@ -132,7 +132,7 @@ ncmpii_new_NC_var(NC_vararray  *vcap,
         /* allocate or expand the space for nameT[key].list */
         if (nameT[key].num % NC_NAME_TABLE_CHUNK == 0)
             nameT[key].list = (int*) NCI_Realloc(nameT[key].list,
-                              (nameT[key].num+NC_NAME_TABLE_CHUNK) * sizeof(int));
+                              (size_t)(nameT[key].num+NC_NAME_TABLE_CHUNK) * sizeof(int));
 
         /* add the new variable ID to the name lookup table
          * the new varid will be vcap->ndefined
@@ -192,7 +192,7 @@ ncmpii_update_name_lookup_table(NC_nametable *nameT,
      */
     if (nameT[key].num % NC_NAME_TABLE_CHUNK == 0)
         nameT[key].list = (int*) NCI_Realloc(nameT[key].list,
-                          (nameT[key].num+NC_NAME_TABLE_CHUNK) * sizeof(int));
+                          (size_t)(nameT[key].num+NC_NAME_TABLE_CHUNK) * sizeof(int));
     nameT[key].list[nameT[key].num] = id;
     nameT[key].num++;
 
@@ -309,9 +309,9 @@ ncmpii_dup_NC_vararray(NC_vararray       *ncap,
         ncap->nameT[i].num = ref->nameT[i].num;
         ncap->nameT[i].list = NULL;
         if (ncap->nameT[i].num > 0) {
-            ncap->nameT[i].list = NCI_Malloc(ncap->nameT[i].num * sizeof(int));
+            ncap->nameT[i].list = NCI_Malloc((size_t)ncap->nameT[i].num * sizeof(int));
             memcpy(ncap->nameT[i].list, ref->nameT[i].list,
-                   ncap->nameT[i].num * sizeof(int));
+                   (size_t)ncap->nameT[i].num * sizeof(int));
         }
     }
 
@@ -728,16 +728,16 @@ err_check:
         if (root_ndims > 0) {
             int root_dimids[NC_MAX_DIMS];
             if (dimids != NULL)
-                memcpy(root_dimids, dimids, root_ndims*sizeof(int));
+                memcpy(root_dimids, dimids, (size_t)root_ndims*sizeof(int));
             else
-                memset(root_dimids, 0, root_ndims*sizeof(int));
+                memset(root_dimids, 0, (size_t)root_ndims*sizeof(int));
             TRACE_COMM(MPI_Bcast)(root_dimids, root_ndims, MPI_INT, 0, ncp->nciop->comm);
             if (mpireturn != MPI_SUCCESS) {
                 if (nname != NULL) free(nname);
                 return ncmpii_handle_error(mpireturn, "MPI_Bcast");
             }
             if (err == NC_NOERR && dimids != NULL &&
-                memcmp(root_dimids, dimids, root_ndims*sizeof(int)))
+                memcmp(root_dimids, dimids, (size_t)root_ndims*sizeof(int)))
                 DEBUG_ASSIGN_ERROR(err, NC_EMULTIDEFINE_VAR_DIMIDS)
         }
